+++ processor/src/test/resources/expected/com/example/repository/Arez_InjectDisabledRepositoryRepository.java
@@ -87,7 +87,7 @@ final class Arez_InjectDisabledRepositoryRepository extends InjectDisabledReposi
 
   @Nonnull
   @Override
-  protected ObservableValue getEntitiesObservableValue() {
+  protected ObservableValue<Stream<InjectDisabledRepository>> getEntitiesObservableValue() {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getEntitiesObservableValue' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
     }