+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithInitializerNameCollisionModelRepository.java
@@ -74,7 +74,8 @@ final class Arez_RepositoryWithInitializerNameCollisionModelRepository extends R
 
   @Nonnull
   @Override
-  protected ObservableValue getEntitiesObservableValue() {
+  protected ObservableValue<Stream<RepositoryWithInitializerNameCollisionModel>> getEntitiesObservableValue(
+      ) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getEntitiesObservableValue' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
     }